story state created can update story as game progesses and story will be updated via story board.
